refactor: [Util.kt] 함수를 용도에 맞는 파일로 분리
